Place all download helper functions into own file.
